change translate session to promise
