+++ server/index.js
@@ -5,7 +5,7 @@ const express = require('express');
 const { Server } = require('socket.io');
 const fs = require("fs")
 const morgan = require("morgan")
-const GoogleTranslateSession = require("./translateEngine/translateSession")
+const translateFile = require("./translateEngine/translateSession")
 const AudioConversion = require("./translateEngine/AudioConversion")
 
 const app = express()
@@ -37,22 +37,10 @@ io.on("connection", (socket) => {
   socket.on('audio', async (data) =>{
     const startTime = Date.now()
     const tempFlacPath = await AudioConversion(data.audioData, './audio/tempM4A.m4a', './audio/serverSaved.flac')
-    
-    var outTxt = ''
-    const GTranslator = new GoogleTranslateSession(data.langSource, data.langTarget, true)
-    const writeStream2Google = GTranslator.googleStream2SocketPlusLog(socket, (finalTranslation) => {
-      return makeSessionRecord (socket.id, data.langSource, data.langTarget, data.audioData, finalTranslation.translation, startTime)
+    translateFile(tempFlacPath, data.langSource, data.langTarget, socket, true).then((finalTranslation) => {
+      console.log(JSON.stringify(finalTranslation))
     })
-    
-    let chunks = [GTranslator.initialRequest]
-    fs.createReadStream(tempFlacPath).on('data', (data) => {
-      chunks.push(GTranslator.requestify(data))
-    }).on('close', () => {
-      chunks.map(chunk => {
-        writeStream2Google.write(chunk)
-      })
-      writeStream2Google.end()
-    })
+    /// transcibeFile()
     
   })
